+++ src/main/java/com/bridgelabz/addressbookapp/repository/AddressBookRepository.java
@@ -3,5 +3,5 @@ package com.bridgelabz.addressbookapp.repository;
 import com.bridgelabz.addressbookapp.entity.AddressBookData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressBookRepository extends JpaRepository<AddressBookData, Integer> {
+public interface AddressBookRepository extends JpaRepository<AddressBookData, Long> {
 }
